Add files via upload
This implementation defines a Node class, which has three attributes: data, next_node and prev_node. The DoublyLinkedList class has head and tail attribute that points to the first and last node in the list respectively. The class also has methods to insert a node at the start, insert a node at the end, delete a node by value, and print the list.

In the example usage, we create a new DoublyLinkedList object and use the insert methods to add some nodes. Then we call the print_list() method to display the contents of the list, and finally we delete a node by value and print the list again. Note that, doubly linked list have the advantage of traversing in both forward and backward direction as it has pointer to both