Add files via upload
In this example, the CircularDoublyLinkedList class has two methods: insert_at_start and delete_by_value. The insert_at_start method creates a new node with the given data, links it to the current head of the list, and makes it the new head. The delete_by_value method searches for a node with the given data and removes it from the list, updating the links of the surrounding nodes accordingly.

It is a basic implementation of circular doubly linked list and you can add more functionality as per your requirement.